Code review changes done , Added translation functionality,Editing functionality
