+++ cypress/e2e/rubriq/BasicTest/basicRubriq.cy.js
@@ -1,0 +1,6 @@
+
+describe('Navigation', function () {
+    it('open  rubriq login page', function() {
+        cy.visit(Cypress.config('baseUrl'))
+    }) 
+})
