--- cypress/e2e/rubriq/BasicTest/basicRubriq.js
@@ -1,6 +1,0 @@
-
-describe('Navigation', function () {
-    it('open  rubriq login page', function() {
-        cy.visit('https://secure-aje.staging.sqr.io/en/rubriq')
-    }) 
-})
