--- cypress/e2e/rubriq/pageObjectRubriq.json
@@ -1,80 +1,0 @@
- {
-    "login": {
-    "email": "[name='email']",
-    "password": "[name='password']" ,
-    "submit": "[type='submit']",
-    "regNewLink": ".skip-tracking.trackable-link",
-    "registrationForm":"[data-cy='registration-form']",
-    "registerbtn":"[data-cy='submit']",
-    "firstName":"[name='firstName']",
-    "lastName":"[name='lastName']",
-    "termsOfService":"#agreeToTOS"
-    },
-    "tabNavigation":{
-    "myAccount":".userAccount",
-    "rubriqTab": "[href='/en/rubriq']",
-    "translateTab": "[title='curie-translation']",
-    "abstractTab": "[title='curie.abstracts']",
-    "editingTab": "[title='editing']",
-    "rubriqPop": ".btn.btn-primary.btn-squared",
-    "ctnBtn": "[title='continue']",
-    "Gotit":"[aria-label='Got it']"
-    },
-    "editing":{
-    "checkDocumentEdit": "[title='check document']",
-    "upgradeNow": "[title='upgrade-now']",
-    "styleOfEngAE": "[id='styleOfEnglish0']",
-    "styleOfEngBE": "[id='styleOfEnglish1']",
-    "errorUploadFile": "[title='file-upload-error']",
-    "strBtn": "[title='start']",
-    "fileUpload": "[title='rubriq-file-upload-text']",
-    "noPlanContentBox": "[title='no-plan-content']",
-    "gotItBtn": "[title='got-it']",
-    "getProfEditing": "[title='curie.get-premium']",
-    "deleteButtonEdit": "[data-icon='trash-can']",
-    "DeleteConfirm": "[title='de-delete-file-yes-text']",
-    "DeleteToastMessage": "[title='%(fileName)s deleted']",
-    "getPremiumBtn": "[title='rubriq.get-premium']",
-    "DownloadBtnEdit": "[data-icon='file-arrow-down']"
-    },
-    "planAndPayment":{
-    "getPremium": "[title='curie.curie-premium']",
-    "getPremiumWithTranslation": "title='rubriq.rubriq-premium-with-translation'",
-    "checkOutWithPaddle": ".paddle-button",
-    "zipCode": "[name='postcode']",
-    "paymentEmail": "[data-testid='authenticationEmailInput']",
-    "countryPayment": "[data-testid='countriesSelect']",
-    "paymentContinue": "[data-testid='combinedAuthenticationLocationFormSubmitButton']",
-    "cardNumberObject": "[data-testid='cardNumberInput']",
-    "cardHolderName": "[data-testid='cardholderNameInput']",
-    "expDateObject": "[data-testid='expiryDateField']",
-    "cvvObject": "[data-testid='cardVerificationValueInput']",
-    "subscribeBtn": "[data-testid='cardPaymentFormSubmitButton']",
-    "plan": "[name='plan']",
-    "myAccountBtn": "[title='my account'] ",
-    "myPlanBtn": "[title='my-plan']",
-    "activeLabel": "[title='active']",
-    "displayedDate":"span[title='Set to auto-renew on %(monthName)s %(day)s, %(year)s']",
-    "cancelSubBtn":"([type='submit'])[2]",
-    "moreDropDown":".rubriqDropdownButton",
-    "cancelBtn": "[title='Cancel']",
-    "cancelPopUpMessage": "[title='rubriq.plan-update-cancel-success']",
-    "paddleFrame": "[name='paddle_frame']",
-    "postCode": "input#postcode",
-    "paddleSubmitBtn": "button.sc-jrQzAO.jIEVBe.sc-kDTinF.jSvKDy",
-    "cardNumberTextBox": "input#cardNumber",
-    "expiryPageObject": "input#expiry",
-    "cvvPageObject": "input#cvv",
-    "cardPaymentFormSubmitBtn": "button[data-testid='cardPaymentFormSubmitButton']",
-    "cancelPlan": ".sc-jrQzAO.jIEVBe.sc-kDTinF.kriCKy"
-    },
-    "translate":{
-    "translateDoc": "[title='translate document']",
-    "chineseLang": "[name='language'][value='zh_CN']",
-    "spanishLang": "[name='language'][value='es_ES']",
-    "portugueseLang": "[name='language'][value='pt_BR']",
-    "dataIconTranslate": "[data-icon='file-word']",
-    "noTranslationPlanContent": "[title='no-plan-translation-content']",
-    "translationLimitExceeded": "[title='number-of-translations-exceeded-title']"
-    }   
-}
