+++ cypress/support/e2e.js
@@ -15,6 +15,7 @@
 
 // Import commands.js using ES2015 syntax:
 import './rubriq_commands'
+import 'cypress-mochawesome-reporter/register';
 
 
 Cypress.on('uncaught:exception', (err, runnable) => {
