+++ package.json
@@ -1,5 +1,10 @@
 {
   "devDependencies": {
-    "cypress": "^14.0.2"
+    "@types/cypress": "^1.1.6",
+    "cypress": "^14.0.2",
+    "cypress-mochawesome-reporter": "^3.8.2",
+    "eslint": "^9.25.1",
+    "eslint-plugin-cypress": "^4.3.0",
+    "fs-extra": "^11.3.0"
   }
 }
